# Optimize `software_titles` query to use indexes