--- a/changes/25160-optimize-software-during-enrollment
+++ b/changes/25160-optimize-software-during-enrollment
@@ -1,0 +1,2 @@
+* Optimized software ingestion queries to use existing DB indexes in the software titles table.
+* Fixed a bug "software not found for checksum" in software ingestion transaction retries.