@lucasmrod 2025-01-23
> Very nice performance improvement!
> 
> Just curious what the numbers looked like if you happened to compare the performance of this server-side approach vs. doing de-duplication in MySQL by splitting the query into separate selects with a `UNION DISTINCT` or `JOIN` strategy. If not, no need to go back and do so. I'm just asking for my own information :)

I haven't. I went straight to the server-side approach (simpler to me).
Only different I can think of is (guess) that 2 queries increases CPU time on the fleet server but reduces "read lock" duration on MySQL, and, simpler/smaller queries are usually better if you don't really care about real-time, which is the case here ("distributed/write" latency).