@gillespi314 2025-01-23
> I went straight to the server-side approach (simpler to me). 

Makes total sense to me, I would have done the same!